bot/requirements.txt
botbuilder-core==4.16.2
botbuilder-integration-aiohttp==4.16.2
aiohttp==3.10.5
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
